move data file to static folder elements
