+++ src/redux/actionsType.js
@@ -3,6 +3,8 @@ export const CHANGE_FILTER = 'CHANGE_FILTER';
 export const FETCH_TOURS_REQUEST = 'FETCH_TOURS_REQUEST';
 export const FETCH_TOURS_SUCCESS = 'FETCH_TOURS_SUCCESS';
 export const FETCH_TOURS_FAILURE = 'FETCH_TOURS_FAILURE';
+export const FETCH_TOUR_DETAIL_SUCCESS = 'FETCH_TOUR_DETAIL_SUCCESS';
+export const FETCH_TOUR_DETAIL_REQUEST = 'FETCH_TOUR_DETAIL_REQUEST';
 
 export const ORDER_TOURS_BY_PRICE_DESC = 'ORDER_TOURS_BY_PRICE_DESC';
 export const ORDER_TOURS_BY_PRICE_ASC = 'ORDER_TOURS_BY_PRICE_ASC';
