--- src/static_data/cities.js
@@ -1,2116 +1,0 @@
-function CitiesData(termSearch) {
-  const list = [
-    {
-      city: 'Tokyo',
-      city_ascii: 'Tokyo',
-      lat: 35.685,
-      lng: 139.7514,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Tōkyō',
-      capital: 'primary',
-      population: 35676000,
-      id: 1392685764,
-    },
-    {
-      city: 'New York',
-      city_ascii: 'New York',
-      lat: 40.6943,
-      lng: -73.9249,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'New York',
-      capital: '',
-      population: 19354922,
-      id: 1840034016,
-    },
-    {
-      city: 'Mexico City',
-      city_ascii: 'Mexico City',
-      lat: 19.4424,
-      lng: -99.131,
-      country: 'Mexico',
-      iso2: 'MX',
-      iso3: 'MEX',
-      admin_name: 'Ciudad de México',
-      capital: 'primary',
-      population: 19028000,
-      id: 1484247881,
-    },
-    {
-      city: 'Mumbai',
-      city_ascii: 'Mumbai',
-      lat: 19.017,
-      lng: 72.857,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Mahārāshtra',
-      capital: 'admin',
-      population: 18978000,
-      id: 1356226629,
-    },
-    {
-      city: 'São Paulo',
-      city_ascii: 'Sao Paulo',
-      lat: -23.5587,
-      lng: -46.625,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'São Paulo',
-      capital: 'admin',
-      population: 18845000,
-      id: 1076532519,
-    },
-    {
-      city: 'Delhi',
-      city_ascii: 'Delhi',
-      lat: 28.67,
-      lng: 77.23,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Delhi',
-      capital: 'admin',
-      population: 15926000,
-      id: 1356872604,
-    },
-    {
-      city: 'Shanghai',
-      city_ascii: 'Shanghai',
-      lat: 31.2165,
-      lng: 121.4365,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shanghai',
-      capital: 'admin',
-      population: 14987000,
-      id: 1156073548,
-    },
-    {
-      city: 'Kolkata',
-      city_ascii: 'Kolkata',
-      lat: 22.495,
-      lng: 88.3247,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'West Bengal',
-      capital: 'admin',
-      population: 14787000,
-      id: 1356060520,
-    },
-    {
-      city: 'Los Angeles',
-      city_ascii: 'Los Angeles',
-      lat: 34.1139,
-      lng: -118.4068,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'California',
-      capital: '',
-      population: 12815475,
-      id: 1840020491,
-    },
-    {
-      city: 'Dhaka',
-      city_ascii: 'Dhaka',
-      lat: 23.7231,
-      lng: 90.4086,
-      country: 'Bangladesh',
-      iso2: 'BD',
-      iso3: 'BGD',
-      admin_name: 'Dhaka',
-      capital: 'primary',
-      population: 12797394,
-      id: 1050529279,
-    },
-    {
-      city: 'Buenos Aires',
-      city_ascii: 'Buenos Aires',
-      lat: -34.6025,
-      lng: -58.3975,
-      country: 'Argentina',
-      iso2: 'AR',
-      iso3: 'ARG',
-      admin_name: 'Buenos Aires, Ciudad Autónoma de',
-      capital: 'primary',
-      population: 12795000,
-      id: 1032717330,
-    },
-    {
-      city: 'Karachi',
-      city_ascii: 'Karachi',
-      lat: 24.87,
-      lng: 66.99,
-      country: 'Pakistan',
-      iso2: 'PK',
-      iso3: 'PAK',
-      admin_name: 'Sindh',
-      capital: 'admin',
-      population: 12130000,
-      id: 1586129469,
-    },
-    {
-      city: 'Cairo',
-      city_ascii: 'Cairo',
-      lat: 30.05,
-      lng: 31.25,
-      country: 'Egypt',
-      iso2: 'EG',
-      iso3: 'EGY',
-      admin_name: 'Al Qāhirah',
-      capital: 'primary',
-      population: 11893000,
-      id: 1818253931,
-    },
-    {
-      city: 'Rio de Janeiro',
-      city_ascii: 'Rio de Janeiro',
-      lat: -22.925,
-      lng: -43.225,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Rio de Janeiro',
-      capital: 'admin',
-      population: 11748000,
-      id: 1076887657,
-    },
-    {
-      city: 'Ōsaka',
-      city_ascii: 'Osaka',
-      lat: 34.75,
-      lng: 135.4601,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Ōsaka',
-      capital: 'admin',
-      population: 11294000,
-      id: 1392419823,
-    },
-    {
-      city: 'Beijing',
-      city_ascii: 'Beijing',
-      lat: 39.9289,
-      lng: 116.3883,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Beijing',
-      capital: 'primary',
-      population: 11106000,
-      id: 1156228865,
-    },
-    {
-      city: 'Manila',
-      city_ascii: 'Manila',
-      lat: 14.6042,
-      lng: 120.9822,
-      country: 'Philippines',
-      iso2: 'PH',
-      iso3: 'PHL',
-      admin_name: 'Manila',
-      capital: 'primary',
-      population: 11100000,
-      id: 1608618140,
-    },
-    {
-      city: 'Moscow',
-      city_ascii: 'Moscow',
-      lat: 55.7522,
-      lng: 37.6155,
-      country: 'Russia',
-      iso2: 'RU',
-      iso3: 'RUS',
-      admin_name: 'Moskva',
-      capital: 'primary',
-      population: 10452000,
-      id: 1643318494,
-    },
-    {
-      city: 'Istanbul',
-      city_ascii: 'Istanbul',
-      lat: 41.105,
-      lng: 29.01,
-      country: 'Turkey',
-      iso2: 'TR',
-      iso3: 'TUR',
-      admin_name: 'İstanbul',
-      capital: 'admin',
-      population: 10061000,
-      id: 1792756324,
-    },
-    {
-      city: 'Paris',
-      city_ascii: 'Paris',
-      lat: 48.8667,
-      lng: 2.3333,
-      country: 'France',
-      iso2: 'FR',
-      iso3: 'FRA',
-      admin_name: 'Île-de-France',
-      capital: 'primary',
-      population: 9904000,
-      id: 1250015082,
-    },
-    {
-      city: 'Seoul',
-      city_ascii: 'Seoul',
-      lat: 37.5663,
-      lng: 126.9997,
-      country: 'Korea, South',
-      iso2: 'KR',
-      iso3: 'KOR',
-      admin_name: 'Seoul',
-      capital: 'primary',
-      population: 9796000,
-      id: 1410836482,
-    },
-    {
-      city: 'Lagos',
-      city_ascii: 'Lagos',
-      lat: 6.4433,
-      lng: 3.3915,
-      country: 'Nigeria',
-      iso2: 'NG',
-      iso3: 'NGA',
-      admin_name: 'Lagos',
-      capital: 'minor',
-      population: 9466000,
-      id: 1566593751,
-    },
-    {
-      city: 'Jakarta',
-      city_ascii: 'Jakarta',
-      lat: -6.1744,
-      lng: 106.8294,
-      country: 'Indonesia',
-      iso2: 'ID',
-      iso3: 'IDN',
-      admin_name: 'Jakarta',
-      capital: 'primary',
-      population: 9125000,
-      id: 1360771077,
-    },
-    {
-      city: 'Guangzhou',
-      city_ascii: 'Guangzhou',
-      lat: 23.145,
-      lng: 113.325,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Guangdong',
-      capital: 'admin',
-      population: 8829000,
-      id: 1156237133,
-    },
-    {
-      city: 'Chicago',
-      city_ascii: 'Chicago',
-      lat: 41.8373,
-      lng: -87.6862,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Illinois',
-      capital: '',
-      population: 8675982,
-      id: 1840000494,
-    },
-    {
-      city: 'London',
-      city_ascii: 'London',
-      lat: 51.5,
-      lng: -0.1167,
-      country: 'United Kingdom',
-      iso2: 'GB',
-      iso3: 'GBR',
-      admin_name: 'London, City of',
-      capital: 'primary',
-      population: 8567000,
-      id: 1826645935,
-    },
-    {
-      city: 'Lima',
-      city_ascii: 'Lima',
-      lat: -12.048,
-      lng: -77.0501,
-      country: 'Peru',
-      iso2: 'PE',
-      iso3: 'PER',
-      admin_name: 'Lima',
-      capital: 'primary',
-      population: 8012000,
-      id: 1604728603,
-    },
-    {
-      city: 'Tehran',
-      city_ascii: 'Tehran',
-      lat: 35.6719,
-      lng: 51.4243,
-      country: 'Iran',
-      iso2: 'IR',
-      iso3: 'IRN',
-      admin_name: 'Tehrān',
-      capital: 'primary',
-      population: 7873000,
-      id: 1364305026,
-    },
-    {
-      city: 'Kinshasa',
-      city_ascii: 'Kinshasa',
-      lat: -4.3297,
-      lng: 15.315,
-      country: 'Congo (Kinshasa)',
-      iso2: 'CD',
-      iso3: 'COD',
-      admin_name: 'Kinshasa',
-      capital: 'primary',
-      population: 7843000,
-      id: 1180000363,
-    },
-    {
-      city: 'Bogotá',
-      city_ascii: 'Bogota',
-      lat: 4.5964,
-      lng: -74.0833,
-      country: 'Colombia',
-      iso2: 'CO',
-      iso3: 'COL',
-      admin_name: 'Bogotá',
-      capital: 'primary',
-      population: 7772000,
-      id: 1170483426,
-    },
-    {
-      city: 'Shenzhen',
-      city_ascii: 'Shenzhen',
-      lat: 22.5524,
-      lng: 114.1221,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Guangdong',
-      capital: 'minor',
-      population: 7581000,
-      id: 1156158707,
-    },
-    {
-      city: 'Wuhan',
-      city_ascii: 'Wuhan',
-      lat: 30.58,
-      lng: 114.27,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Hubei',
-      capital: 'admin',
-      population: 7243000,
-      id: 1156117184,
-    },
-    {
-      city: 'Hong Kong',
-      city_ascii: 'Hong Kong',
-      lat: 22.305,
-      lng: 114.185,
-      country: 'Hong Kong',
-      iso2: 'HK',
-      iso3: 'HKG',
-      admin_name: '',
-      capital: '',
-      population: 7206000,
-      id: 1344982653,
-    },
-    {
-      city: 'Tianjin',
-      city_ascii: 'Tianjin',
-      lat: 39.13,
-      lng: 117.2,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Tianjin',
-      capital: 'admin',
-      population: 7180000,
-      id: 1156174046,
-    },
-    {
-      city: 'Chennai',
-      city_ascii: 'Chennai',
-      lat: 13.09,
-      lng: 80.28,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Tamil Nādu ',
-      capital: 'admin',
-      population: 7163000,
-      id: 1356374944,
-    },
-    {
-      city: 'Taipei',
-      city_ascii: 'Taipei',
-      lat: 25.0358,
-      lng: 121.5683,
-      country: 'Taiwan',
-      iso2: 'TW',
-      iso3: 'TWN',
-      admin_name: 'Taipei',
-      capital: 'primary',
-      population: 6900273,
-      id: 1158881289,
-    },
-    {
-      city: 'Bengalūru',
-      city_ascii: 'Bengaluru',
-      lat: 12.97,
-      lng: 77.56,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Karnātaka',
-      capital: 'admin',
-      population: 6787000,
-      id: 1356410365,
-    },
-    {
-      city: 'Bangkok',
-      city_ascii: 'Bangkok',
-      lat: 13.75,
-      lng: 100.5166,
-      country: 'Thailand',
-      iso2: 'TH',
-      iso3: 'THA',
-      admin_name: 'Krung Thep Maha Nakhon',
-      capital: 'primary',
-      population: 6704000,
-      id: 1764068610,
-    },
-    {
-      city: 'Lahore',
-      city_ascii: 'Lahore',
-      lat: 31.56,
-      lng: 74.35,
-      country: 'Pakistan',
-      iso2: 'PK',
-      iso3: 'PAK',
-      admin_name: 'Punjab',
-      capital: 'admin',
-      population: 6577000,
-      id: 1586801463,
-    },
-    {
-      city: 'Chongqing',
-      city_ascii: 'Chongqing',
-      lat: 29.565,
-      lng: 106.595,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Chongqing',
-      capital: 'admin',
-      population: 6461000,
-      id: 1156936556,
-    },
-    {
-      city: 'Miami',
-      city_ascii: 'Miami',
-      lat: 25.7839,
-      lng: -80.2102,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Florida',
-      capital: '',
-      population: 6381966,
-      id: 1840015149,
-    },
-    {
-      city: 'Hyderabad',
-      city_ascii: 'Hyderabad',
-      lat: 17.4,
-      lng: 78.48,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Telangana',
-      capital: 'admin',
-      population: 6376000,
-      id: 1356871768,
-    },
-    {
-      city: 'Dallas',
-      city_ascii: 'Dallas',
-      lat: 32.7936,
-      lng: -96.7662,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Texas',
-      capital: '',
-      population: 5733259,
-      id: 1840019440,
-    },
-    {
-      city: 'Santiago',
-      city_ascii: 'Santiago',
-      lat: -33.45,
-      lng: -70.667,
-      country: 'Chile',
-      iso2: 'CL',
-      iso3: 'CHL',
-      admin_name: 'Región Metropolitana',
-      capital: 'primary',
-      population: 5720000,
-      id: 1152554349,
-    },
-    {
-      city: 'Philadelphia',
-      city_ascii: 'Philadelphia',
-      lat: 40.0077,
-      lng: -75.1339,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Pennsylvania',
-      capital: '',
-      population: 5637884,
-      id: 1840000673,
-    },
-    {
-      city: 'Belo Horizonte',
-      city_ascii: 'Belo Horizonte',
-      lat: -19.915,
-      lng: -43.915,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Minas Gerais',
-      capital: 'admin',
-      population: 5575000,
-      id: 1076967355,
-    },
-    {
-      city: 'Madrid',
-      city_ascii: 'Madrid',
-      lat: 40.4,
-      lng: -3.6834,
-      country: 'Spain',
-      iso2: 'ES',
-      iso3: 'ESP',
-      admin_name: 'Madrid',
-      capital: 'primary',
-      population: 5567000,
-      id: 1724616994,
-    },
-    {
-      city: 'Houston',
-      city_ascii: 'Houston',
-      lat: 29.7869,
-      lng: -95.3905,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Texas',
-      capital: '',
-      population: 5446468,
-      id: 1840020925,
-    },
-    {
-      city: 'Ahmadābād',
-      city_ascii: 'Ahmadabad',
-      lat: 23.0301,
-      lng: 72.58,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Gujarāt',
-      capital: 'minor',
-      population: 5375000,
-      id: 1356304381,
-    },
-    {
-      city: 'Ho Chi Minh City',
-      city_ascii: 'Ho Chi Minh City',
-      lat: 10.78,
-      lng: 106.695,
-      country: 'Vietnam',
-      iso2: 'VN',
-      iso3: 'VNM',
-      admin_name: 'Hồ Chí Minh',
-      capital: 'admin',
-      population: 5314000,
-      id: 1704774326,
-    },
-    {
-      city: 'Washington',
-      city_ascii: 'Washington',
-      lat: 38.9047,
-      lng: -77.0163,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'District of Columbia',
-      capital: 'primary',
-      population: 5289420,
-      id: 1840006060,
-    },
-    {
-      city: 'Atlanta',
-      city_ascii: 'Atlanta',
-      lat: 33.7627,
-      lng: -84.4225,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Georgia',
-      capital: 'admin',
-      population: 5228750,
-      id: 1840013660,
-    },
-    {
-      city: 'Toronto',
-      city_ascii: 'Toronto',
-      lat: 43.7,
-      lng: -79.42,
-      country: 'Canada',
-      iso2: 'CA',
-      iso3: 'CAN',
-      admin_name: 'Ontario',
-      capital: 'admin',
-      population: 5213000,
-      id: 1124279679,
-    },
-    {
-      city: 'Singapore',
-      city_ascii: 'Singapore',
-      lat: 1.293,
-      lng: 103.8558,
-      country: 'Singapore',
-      iso2: 'SG',
-      iso3: 'SGP',
-      admin_name: '',
-      capital: 'primary',
-      population: 5183700,
-      id: 1702341327,
-    },
-    {
-      city: 'Luanda',
-      city_ascii: 'Luanda',
-      lat: -8.8383,
-      lng: 13.2344,
-      country: 'Angola',
-      iso2: 'AO',
-      iso3: 'AGO',
-      admin_name: 'Luanda',
-      capital: 'primary',
-      population: 5172900,
-      id: 1024949724,
-    },
-    {
-      city: 'Baghdad',
-      city_ascii: 'Baghdad',
-      lat: 33.3386,
-      lng: 44.3939,
-      country: 'Iraq',
-      iso2: 'IQ',
-      iso3: 'IRQ',
-      admin_name: 'Baghdād',
-      capital: 'primary',
-      population: 5054000,
-      id: 1368596238,
-    },
-    {
-      city: 'Barcelona',
-      city_ascii: 'Barcelona',
-      lat: 41.3833,
-      lng: 2.1834,
-      country: 'Spain',
-      iso2: 'ES',
-      iso3: 'ESP',
-      admin_name: 'Catalonia',
-      capital: 'admin',
-      population: 4920000,
-      id: 1724594040,
-    },
-    {
-      city: 'Hāora',
-      city_ascii: 'Haora',
-      lat: 22.5804,
-      lng: 88.3299,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'West Bengal',
-      capital: '',
-      population: 4841638,
-      id: 1356986516,
-    },
-    {
-      city: 'Shenyang',
-      city_ascii: 'Shenyang',
-      lat: 41.805,
-      lng: 123.45,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Liaoning',
-      capital: 'admin',
-      population: 4787000,
-      id: 1156309382,
-    },
-    {
-      city: 'Khartoum',
-      city_ascii: 'Khartoum',
-      lat: 15.5881,
-      lng: 32.5342,
-      country: 'Sudan',
-      iso2: 'SD',
-      iso3: 'SDN',
-      admin_name: 'Khartoum',
-      capital: 'primary',
-      population: 4754000,
-      id: 1729268475,
-    },
-    {
-      city: 'Pune',
-      city_ascii: 'Pune',
-      lat: 18.53,
-      lng: 73.85,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Mahārāshtra',
-      capital: '',
-      population: 4672000,
-      id: 1356081074,
-    },
-    {
-      city: 'Boston',
-      city_ascii: 'Boston',
-      lat: 42.3188,
-      lng: -71.0846,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Massachusetts',
-      capital: 'admin',
-      population: 4637537,
-      id: 1840000455,
-    },
-    {
-      city: 'Sydney',
-      city_ascii: 'Sydney',
-      lat: -33.92,
-      lng: 151.1852,
-      country: 'Australia',
-      iso2: 'AU',
-      iso3: 'AUS',
-      admin_name: 'New South Wales',
-      capital: 'admin',
-      population: 4630000,
-      id: 1036074917,
-    },
-    {
-      city: 'Saint Petersburg',
-      city_ascii: 'Saint Petersburg',
-      lat: 59.939,
-      lng: 30.316,
-      country: 'Russia',
-      iso2: 'RU',
-      iso3: 'RUS',
-      admin_name: 'Sankt-Peterburg',
-      capital: 'admin',
-      population: 4553000,
-      id: 1643616350,
-    },
-    {
-      city: 'Chittagong',
-      city_ascii: 'Chittagong',
-      lat: 22.33,
-      lng: 91.8,
-      country: 'Bangladesh',
-      iso2: 'BD',
-      iso3: 'BGD',
-      admin_name: 'Chittagong',
-      capital: 'admin',
-      population: 4529000,
-      id: 1050830722,
-    },
-    {
-      city: 'Dongguan',
-      city_ascii: 'Dongguan',
-      lat: 23.0489,
-      lng: 113.7447,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Guangdong',
-      capital: 'minor',
-      population: 4528000,
-      id: 1156478242,
-    },
-    {
-      city: 'Riyadh',
-      city_ascii: 'Riyadh',
-      lat: 24.6408,
-      lng: 46.7727,
-      country: 'Saudi Arabia',
-      iso2: 'SA',
-      iso3: 'SAU',
-      admin_name: 'Ar Riyāḑ',
-      capital: 'primary',
-      population: 4465000,
-      id: 1682999334,
-    },
-    {
-      city: 'Hanoi',
-      city_ascii: 'Hanoi',
-      lat: 21.0333,
-      lng: 105.85,
-      country: 'Vietnam',
-      iso2: 'VN',
-      iso3: 'VNM',
-      admin_name: 'Hà Nội',
-      capital: 'primary',
-      population: 4378000,
-      id: 1704413791,
-    },
-    {
-      city: 'Guadalajara',
-      city_ascii: 'Guadalajara',
-      lat: 20.67,
-      lng: -103.33,
-      country: 'Mexico',
-      iso2: 'MX',
-      iso3: 'MEX',
-      admin_name: 'Jalisco',
-      capital: 'admin',
-      population: 4198000,
-      id: 1484950208,
-    },
-    {
-      city: 'Melbourne',
-      city_ascii: 'Melbourne',
-      lat: -37.82,
-      lng: 144.975,
-      country: 'Australia',
-      iso2: 'AU',
-      iso3: 'AUS',
-      admin_name: 'Victoria',
-      capital: 'admin',
-      population: 4170000,
-      id: 1036533631,
-    },
-    {
-      city: 'Alexandria',
-      city_ascii: 'Alexandria',
-      lat: 31.2,
-      lng: 29.95,
-      country: 'Egypt',
-      iso2: 'EG',
-      iso3: 'EGY',
-      admin_name: 'Al Iskandarīyah',
-      capital: 'admin',
-      population: 4165000,
-      id: 1818695837,
-    },
-    {
-      city: 'Chengdu',
-      city_ascii: 'Chengdu',
-      lat: 30.67,
-      lng: 104.07,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Sichuan',
-      capital: 'admin',
-      population: 4123000,
-      id: 1156421555,
-    },
-    {
-      city: 'Rangoon',
-      city_ascii: 'Rangoon',
-      lat: 16.7834,
-      lng: 96.1667,
-      country: 'Burma',
-      iso2: 'MM',
-      iso3: 'MMR',
-      admin_name: 'Yangon',
-      capital: 'primary',
-      population: 4088000,
-      id: 1104616656,
-    },
-    {
-      city: 'Phoenix',
-      city_ascii: 'Phoenix',
-      lat: 33.5722,
-      lng: -112.0891,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Arizona',
-      capital: 'admin',
-      population: 4081849,
-      id: 1840020568,
-    },
-    {
-      city: 'Xi’an',
-      city_ascii: "Xi'an",
-      lat: 34.275,
-      lng: 108.895,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shaanxi',
-      capital: 'admin',
-      population: 4009000,
-      id: 1156244079,
-    },
-    {
-      city: 'Porto Alegre',
-      city_ascii: 'Porto Alegre',
-      lat: -30.05,
-      lng: -51.2,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Rio Grande do Sul',
-      capital: 'admin',
-      population: 3917000,
-      id: 1076658221,
-    },
-    {
-      city: 'Sūrat',
-      city_ascii: 'Surat',
-      lat: 21.2,
-      lng: 72.84,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Gujarāt',
-      capital: '',
-      population: 3842000,
-      id: 1356758738,
-    },
-    {
-      city: 'Hechi',
-      city_ascii: 'Hechi',
-      lat: 23.0965,
-      lng: 109.6091,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Guangxi',
-      capital: '',
-      population: 3830000,
-      id: 1156868701,
-    },
-    {
-      city: 'Abidjan',
-      city_ascii: 'Abidjan',
-      lat: 5.32,
-      lng: -4.04,
-      country: 'Côte D’Ivoire',
-      iso2: 'CI',
-      iso3: 'CIV',
-      admin_name: 'Abidjan',
-      capital: 'primary',
-      population: 3802000,
-      id: 1384207980,
-    },
-    {
-      city: 'Brasília',
-      city_ascii: 'Brasilia',
-      lat: -15.7833,
-      lng: -47.9161,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Distrito Federal',
-      capital: 'primary',
-      population: 3716996,
-      id: 1076144436,
-    },
-    {
-      city: 'Ankara',
-      city_ascii: 'Ankara',
-      lat: 39.9272,
-      lng: 32.8644,
-      country: 'Turkey',
-      iso2: 'TR',
-      iso3: 'TUR',
-      admin_name: 'Ankara',
-      capital: 'primary',
-      population: 3716000,
-      id: 1792572891,
-    },
-    {
-      city: 'Monterrey',
-      city_ascii: 'Monterrey',
-      lat: 25.67,
-      lng: -100.33,
-      country: 'Mexico',
-      iso2: 'MX',
-      iso3: 'MEX',
-      admin_name: 'Nuevo León',
-      capital: 'admin',
-      population: 3712000,
-      id: 1484559591,
-    },
-    {
-      city: 'Yokohama',
-      city_ascii: 'Yokohama',
-      lat: 35.32,
-      lng: 139.58,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Kanagawa',
-      capital: 'admin',
-      population: 3697894,
-      id: 1392118339,
-    },
-    {
-      city: 'Nanjing',
-      city_ascii: 'Nanjing',
-      lat: 32.05,
-      lng: 118.78,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Jiangsu',
-      capital: 'admin',
-      population: 3679000,
-      id: 1156644065,
-    },
-    {
-      city: 'Montréal',
-      city_ascii: 'Montreal',
-      lat: 45.5,
-      lng: -73.5833,
-      country: 'Canada',
-      iso2: 'CA',
-      iso3: 'CAN',
-      admin_name: 'Québec',
-      capital: '',
-      population: 3678000,
-      id: 1124586170,
-    },
-    {
-      city: 'Guiyang',
-      city_ascii: 'Guiyang',
-      lat: 26.58,
-      lng: 106.72,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Guizhou',
-      capital: 'admin',
-      population: 3662000,
-      id: 1156932620,
-    },
-    {
-      city: 'Recife',
-      city_ascii: 'Recife',
-      lat: -8.0756,
-      lng: -34.9156,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Pernambuco',
-      capital: 'admin',
-      population: 3651000,
-      id: 1076137337,
-    },
-    {
-      city: 'Seattle',
-      city_ascii: 'Seattle',
-      lat: 47.6211,
-      lng: -122.3244,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Washington',
-      capital: '',
-      population: 3643765,
-      id: 1840021117,
-    },
-    {
-      city: 'Harbin',
-      city_ascii: 'Harbin',
-      lat: 45.75,
-      lng: 126.65,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Heilongjiang',
-      capital: 'admin',
-      population: 3621000,
-      id: 1156241528,
-    },
-    {
-      city: 'San Francisco',
-      city_ascii: 'San Francisco',
-      lat: 37.7562,
-      lng: -122.443,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'California',
-      capital: '',
-      population: 3603761,
-      id: 1840021543,
-    },
-    {
-      city: 'Fortaleza',
-      city_ascii: 'Fortaleza',
-      lat: -3.75,
-      lng: -38.58,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Ceará',
-      capital: 'admin',
-      population: 3602319,
-      id: 1076567885,
-    },
-    {
-      city: 'Zhangzhou',
-      city_ascii: 'Zhangzhou',
-      lat: 24.5204,
-      lng: 117.67,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Fujian',
-      capital: 'minor',
-      population: 3531147,
-      id: 1156241637,
-    },
-    {
-      city: 'Detroit',
-      city_ascii: 'Detroit',
-      lat: 42.3834,
-      lng: -83.1024,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Michigan',
-      capital: '',
-      population: 3522206,
-      id: 1840003971,
-    },
-    {
-      city: 'Salvador',
-      city_ascii: 'Salvador',
-      lat: -12.97,
-      lng: -38.48,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Bahia',
-      capital: 'admin',
-      population: 3484000,
-      id: 1076923789,
-    },
-    {
-      city: 'Busan',
-      city_ascii: 'Busan',
-      lat: 35.0951,
-      lng: 129.01,
-      country: 'Korea, South',
-      iso2: 'KR',
-      iso3: 'KOR',
-      admin_name: 'Busan',
-      capital: 'admin',
-      population: 3480000,
-      id: 1410601465,
-    },
-    {
-      city: 'Johannesburg',
-      city_ascii: 'Johannesburg',
-      lat: -26.17,
-      lng: 28.03,
-      country: 'South Africa',
-      iso2: 'ZA',
-      iso3: 'ZAF',
-      admin_name: 'Gauteng',
-      capital: 'admin',
-      population: 3435000,
-      id: 1710550792,
-    },
-    {
-      city: 'Berlin',
-      city_ascii: 'Berlin',
-      lat: 52.5218,
-      lng: 13.4015,
-      country: 'Germany',
-      iso2: 'DE',
-      iso3: 'DEU',
-      admin_name: 'Berlin',
-      capital: 'primary',
-      population: 3406000,
-      id: 1276451290,
-    },
-    {
-      city: 'Algiers',
-      city_ascii: 'Algiers',
-      lat: 36.7631,
-      lng: 3.0506,
-      country: 'Algeria',
-      iso2: 'DZ',
-      iso3: 'DZA',
-      admin_name: 'Alger',
-      capital: 'primary',
-      population: 3354000,
-      id: 1012973369,
-    },
-    {
-      city: 'Rome',
-      city_ascii: 'Rome',
-      lat: 41.896,
-      lng: 12.4833,
-      country: 'Italy',
-      iso2: 'IT',
-      iso3: 'ITA',
-      admin_name: 'Lazio',
-      capital: 'primary',
-      population: 3339000,
-      id: 1380382862,
-    },
-    {
-      city: 'Pyongyang',
-      city_ascii: 'Pyongyang',
-      lat: 39.0194,
-      lng: 125.7547,
-      country: 'Korea, North',
-      iso2: 'KP',
-      iso3: 'PRK',
-      admin_name: 'P’yŏngyang',
-      capital: 'primary',
-      population: 3300000,
-      id: 1408738594,
-    },
-    {
-      city: 'Medellín',
-      city_ascii: 'Medellin',
-      lat: 6.275,
-      lng: -75.575,
-      country: 'Colombia',
-      iso2: 'CO',
-      iso3: 'COL',
-      admin_name: 'Antioquia',
-      capital: 'admin',
-      population: 3297000,
-      id: 1170680389,
-    },
-    {
-      city: 'Kabul',
-      city_ascii: 'Kabul',
-      lat: 34.5167,
-      lng: 69.1833,
-      country: 'Afghanistan',
-      iso2: 'AF',
-      iso3: 'AFG',
-      admin_name: 'Kābul',
-      capital: 'primary',
-      population: 3277000,
-      id: 1004993580,
-    },
-    {
-      city: 'Athens',
-      city_ascii: 'Athens',
-      lat: 37.9833,
-      lng: 23.7333,
-      country: 'Greece',
-      iso2: 'GR',
-      iso3: 'GRC',
-      admin_name: 'Attikí',
-      capital: 'primary',
-      population: 3242000,
-      id: 1300715560,
-    },
-    {
-      city: 'Nagoya',
-      city_ascii: 'Nagoya',
-      lat: 35.155,
-      lng: 136.915,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Aichi',
-      capital: 'admin',
-      population: 3230000,
-      id: 1392407472,
-    },
-    {
-      city: 'Cape Town',
-      city_ascii: 'Cape Town',
-      lat: -33.92,
-      lng: 18.435,
-      country: 'South Africa',
-      iso2: 'ZA',
-      iso3: 'ZAF',
-      admin_name: 'Western Cape',
-      capital: 'primary',
-      population: 3215000,
-      id: 1710680650,
-    },
-    {
-      city: 'San Diego',
-      city_ascii: 'San Diego',
-      lat: 32.8312,
-      lng: -117.1225,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'California',
-      capital: '',
-      population: 3210314,
-      id: 1840021990,
-    },
-    {
-      city: 'Changchun',
-      city_ascii: 'Changchun',
-      lat: 43.865,
-      lng: 125.34,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Jilin',
-      capital: 'admin',
-      population: 3183000,
-      id: 1156078103,
-    },
-    {
-      city: 'Casablanca',
-      city_ascii: 'Casablanca',
-      lat: 33.6,
-      lng: -7.6164,
-      country: 'Morocco',
-      iso2: 'MA',
-      iso3: 'MAR',
-      admin_name: 'Casablanca-Settat',
-      capital: 'admin',
-      population: 3181000,
-      id: 1504175315,
-    },
-    {
-      city: 'Dalian',
-      city_ascii: 'Dalian',
-      lat: 38.9228,
-      lng: 121.6298,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Liaoning',
-      capital: '',
-      population: 3167000,
-      id: 1156175472,
-    },
-    {
-      city: 'Cawnpore',
-      city_ascii: 'Cawnpore',
-      lat: 26.46,
-      lng: 80.32,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Uttar Pradesh',
-      capital: '',
-      population: 3162000,
-      id: 1356575238,
-    },
-    {
-      city: 'Kano',
-      city_ascii: 'Kano',
-      lat: 12,
-      lng: 8.52,
-      country: 'Nigeria',
-      iso2: 'NG',
-      iso3: 'NGA',
-      admin_name: 'Kano',
-      capital: 'admin',
-      population: 3140000,
-      id: 1566422868,
-    },
-    {
-      city: 'Tel Aviv-Yafo',
-      city_ascii: 'Tel Aviv-Yafo',
-      lat: 32.08,
-      lng: 34.77,
-      country: 'Israel',
-      iso2: 'IL',
-      iso3: 'ISR',
-      admin_name: 'Tel Aviv',
-      capital: 'admin',
-      population: 3112000,
-      id: 1376401542,
-    },
-    {
-      city: 'Addis Ababa',
-      city_ascii: 'Addis Ababa',
-      lat: 9.0333,
-      lng: 38.7,
-      country: 'Ethiopia',
-      iso2: 'ET',
-      iso3: 'ETH',
-      admin_name: 'Ādīs Ābeba',
-      capital: 'primary',
-      population: 3100000,
-      id: 1231824991,
-    },
-    {
-      city: 'Curitiba',
-      city_ascii: 'Curitiba',
-      lat: -25.42,
-      lng: -49.32,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'Paraná',
-      capital: 'admin',
-      population: 3084000,
-      id: 1076701712,
-    },
-    {
-      city: 'Zibo',
-      city_ascii: 'Zibo',
-      lat: 36.8,
-      lng: 118.05,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shandong',
-      capital: '',
-      population: 3061000,
-      id: 1156381094,
-    },
-    {
-      city: 'Jeddah',
-      city_ascii: 'Jeddah',
-      lat: 21.5169,
-      lng: 39.2192,
-      country: 'Saudi Arabia',
-      iso2: 'SA',
-      iso3: 'SAU',
-      admin_name: 'Makkah al Mukarramah',
-      capital: '',
-      population: 3012000,
-      id: 1682926944,
-    },
-    {
-      city: 'Nairobi',
-      city_ascii: 'Nairobi',
-      lat: -1.2833,
-      lng: 36.8167,
-      country: 'Kenya',
-      iso2: 'KE',
-      iso3: 'KEN',
-      admin_name: 'Nairobi City',
-      capital: 'primary',
-      population: 3010000,
-      id: 1404000661,
-    },
-    {
-      city: 'Hangzhou',
-      city_ascii: 'Hangzhou',
-      lat: 30.25,
-      lng: 120.17,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Zhejiang',
-      capital: 'admin',
-      population: 3007000,
-      id: 1156275243,
-    },
-    {
-      city: 'Benoni',
-      city_ascii: 'Benoni',
-      lat: -26.1496,
-      lng: 28.3299,
-      country: 'South Africa',
-      iso2: 'ZA',
-      iso3: 'ZAF',
-      admin_name: 'Gauteng',
-      capital: '',
-      population: 2986000,
-      id: 1710880986,
-    },
-    {
-      city: 'Caracas',
-      city_ascii: 'Caracas',
-      lat: 10.501,
-      lng: -66.917,
-      country: 'Venezuela',
-      iso2: 'VE',
-      iso3: 'VEN',
-      admin_name: 'Distrito Capital',
-      capital: 'primary',
-      population: 2985000,
-      id: 1862748204,
-    },
-    {
-      city: 'Milan',
-      city_ascii: 'Milan',
-      lat: 45.47,
-      lng: 9.205,
-      country: 'Italy',
-      iso2: 'IT',
-      iso3: 'ITA',
-      admin_name: 'Lombardy',
-      capital: 'admin',
-      population: 2945000,
-      id: 1380724377,
-    },
-    {
-      city: 'Stuttgart',
-      city_ascii: 'Stuttgart',
-      lat: 48.78,
-      lng: 9.2,
-      country: 'Germany',
-      iso2: 'DE',
-      iso3: 'DEU',
-      admin_name: 'Baden-Württemberg',
-      capital: 'admin',
-      population: 2944700,
-      id: 1276171358,
-    },
-    {
-      city: 'Kunming',
-      city_ascii: 'Kunming',
-      lat: 25.07,
-      lng: 102.68,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Yunnan',
-      capital: 'admin',
-      population: 2931000,
-      id: 1156477539,
-    },
-    {
-      city: 'Dar es Salaam',
-      city_ascii: 'Dar es Salaam',
-      lat: -6.8,
-      lng: 39.2683,
-      country: 'Tanzania',
-      iso2: 'TZ',
-      iso3: 'TZA',
-      admin_name: 'Dar es Salaam',
-      capital: 'primary',
-      population: 2930000,
-      id: 1834843853,
-    },
-    {
-      city: 'Minneapolis',
-      city_ascii: 'Minneapolis',
-      lat: 44.9635,
-      lng: -93.2678,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Minnesota',
-      capital: '',
-      population: 2926757,
-      id: 1840007830,
-    },
-    {
-      city: 'Jaipur',
-      city_ascii: 'Jaipur',
-      lat: 26.9211,
-      lng: 75.81,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Rājasthān',
-      capital: 'admin',
-      population: 2917000,
-      id: 1356293537,
-    },
-    {
-      city: 'Taiyuan',
-      city_ascii: 'Taiyuan',
-      lat: 37.875,
-      lng: 112.5451,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shanxi',
-      capital: 'admin',
-      population: 2913000,
-      id: 1156632014,
-    },
-    {
-      city: 'Frankfurt',
-      city_ascii: 'Frankfurt',
-      lat: 50.1,
-      lng: 8.675,
-      country: 'Germany',
-      iso2: 'DE',
-      iso3: 'DEU',
-      admin_name: 'Hesse',
-      capital: 'minor',
-      population: 2895000,
-      id: 1276054552,
-    },
-    {
-      city: 'Qingdao',
-      city_ascii: 'Qingdao',
-      lat: 36.09,
-      lng: 120.33,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shandong',
-      capital: 'minor',
-      population: 2866000,
-      id: 1156112588,
-    },
-    {
-      city: 'Surabaya',
-      city_ascii: 'Surabaya',
-      lat: -7.2492,
-      lng: 112.7508,
-      country: 'Indonesia',
-      iso2: 'ID',
-      iso3: 'IDN',
-      admin_name: 'Jawa Timur',
-      capital: 'admin',
-      population: 2845000,
-      id: 1360484663,
-    },
-    {
-      city: 'Lisbon',
-      city_ascii: 'Lisbon',
-      lat: 38.7227,
-      lng: -9.1449,
-      country: 'Portugal',
-      iso2: 'PT',
-      iso3: 'PRT',
-      admin_name: 'Lisboa',
-      capital: 'primary',
-      population: 2812000,
-      id: 1620619017,
-    },
-    {
-      city: 'Tampa',
-      city_ascii: 'Tampa',
-      lat: 27.9942,
-      lng: -82.4451,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Florida',
-      capital: '',
-      population: 2804240,
-      id: 1840015982,
-    },
-    {
-      city: 'Jinan',
-      city_ascii: 'Jinan',
-      lat: 36.675,
-      lng: 116.995,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Shandong',
-      capital: 'admin',
-      population: 2798000,
-      id: 1156972469,
-    },
-    {
-      city: 'Fukuoka',
-      city_ascii: 'Fukuoka',
-      lat: 33.595,
-      lng: 130.41,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Fukuoka',
-      capital: 'admin',
-      population: 2792000,
-      id: 1392576294,
-    },
-    {
-      city: 'Campinas',
-      city_ascii: 'Campinas',
-      lat: -22.9,
-      lng: -47.1,
-      country: 'Brazil',
-      iso2: 'BR',
-      iso3: 'BRA',
-      admin_name: 'São Paulo',
-      capital: '',
-      population: 2791000,
-      id: 1076145766,
-    },
-    {
-      city: 'Denver',
-      city_ascii: 'Denver',
-      lat: 39.7621,
-      lng: -104.8759,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'Colorado',
-      capital: 'admin',
-      population: 2787266,
-      id: 1840018789,
-    },
-    {
-      city: 'Kaohsiung',
-      city_ascii: 'Kaohsiung',
-      lat: 22.6333,
-      lng: 120.2666,
-      country: 'Taiwan',
-      iso2: 'TW',
-      iso3: 'TWN',
-      admin_name: 'Kaohsiung',
-      capital: 'admin',
-      population: 2769072,
-      id: 1158331334,
-    },
-    {
-      city: 'Quezon City',
-      city_ascii: 'Quezon City',
-      lat: 14.6504,
-      lng: 121.03,
-      country: 'Philippines',
-      iso2: 'PH',
-      iso3: 'PHL',
-      admin_name: 'Quezon',
-      capital: 'admin',
-      population: 2761720,
-      id: 1608974097,
-    },
-    {
-      city: 'Katowice',
-      city_ascii: 'Katowice',
-      lat: 50.2604,
-      lng: 19.02,
-      country: 'Poland',
-      iso2: 'PL',
-      iso3: 'POL',
-      admin_name: 'Śląskie',
-      capital: 'admin',
-      population: 2746000,
-      id: 1616798495,
-    },
-    {
-      city: 'Aleppo',
-      city_ascii: 'Aleppo',
-      lat: 36.23,
-      lng: 37.17,
-      country: 'Syria',
-      iso2: 'SY',
-      iso3: 'SYR',
-      admin_name: 'Ḩalab',
-      capital: 'admin',
-      population: 2738000,
-      id: 1760366651,
-    },
-    {
-      city: 'Durban',
-      city_ascii: 'Durban',
-      lat: -29.865,
-      lng: 30.98,
-      country: 'South Africa',
-      iso2: 'ZA',
-      iso3: 'ZAF',
-      admin_name: 'KwaZulu-Natal',
-      capital: '',
-      population: 2729000,
-      id: 1710149138,
-    },
-    {
-      city: 'Kyiv',
-      city_ascii: 'Kyiv',
-      lat: 50.4334,
-      lng: 30.5166,
-      country: 'Ukraine',
-      iso2: 'UA',
-      iso3: 'UKR',
-      admin_name: 'Kyyiv, Misto',
-      capital: 'primary',
-      population: 2709000,
-      id: 1804382913,
-    },
-    {
-      city: 'Lucknow',
-      city_ascii: 'Lucknow',
-      lat: 26.855,
-      lng: 80.915,
-      country: 'India',
-      iso2: 'IN',
-      iso3: 'IND',
-      admin_name: 'Uttar Pradesh',
-      capital: 'admin',
-      population: 2695000,
-      id: 1356891790,
-    },
-    {
-      city: 'Giza',
-      city_ascii: 'Giza',
-      lat: 30.01,
-      lng: 31.19,
-      country: 'Egypt',
-      iso2: 'EG',
-      iso3: 'EGY',
-      admin_name: 'Al Jīzah',
-      capital: 'admin',
-      population: 2681863,
-      id: 1818925479,
-    },
-    {
-      city: 'Zhengzhou',
-      city_ascii: 'Zhengzhou',
-      lat: 34.755,
-      lng: 113.6651,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Henan',
-      capital: 'admin',
-      population: 2636000,
-      id: 1156183137,
-    },
-    {
-      city: 'Taichung',
-      city_ascii: 'Taichung',
-      lat: 24.1521,
-      lng: 120.6817,
-      country: 'Taiwan',
-      iso2: 'TW',
-      iso3: 'TWN',
-      admin_name: 'Taichung',
-      capital: 'admin',
-      population: 2629323,
-      id: 1158689622,
-    },
-    {
-      city: 'Brooklyn',
-      city_ascii: 'Brooklyn',
-      lat: 40.6501,
-      lng: -73.9496,
-      country: 'United States',
-      iso2: 'US',
-      iso3: 'USA',
-      admin_name: 'New York',
-      capital: '',
-      population: 2629150,
-      id: 1840034030,
-    },
-    {
-      city: 'Ibadan',
-      city_ascii: 'Ibadan',
-      lat: 7.38,
-      lng: 3.93,
-      country: 'Nigeria',
-      iso2: 'NG',
-      iso3: 'NGA',
-      admin_name: 'Oyo',
-      capital: 'admin',
-      population: 2628000,
-      id: 1566366407,
-    },
-    {
-      city: 'Faisalabad',
-      city_ascii: 'Faisalabad',
-      lat: 31.41,
-      lng: 73.11,
-      country: 'Pakistan',
-      iso2: 'PK',
-      iso3: 'PAK',
-      admin_name: 'Punjab',
-      capital: 'minor',
-      population: 2617000,
-      id: 1586323916,
-    },
-    {
-      city: 'Fuzhou',
-      city_ascii: 'Fuzhou',
-      lat: 26.08,
-      lng: 119.3,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Fujian',
-      capital: 'admin',
-      population: 2606000,
-      id: 1156188037,
-    },
-    {
-      city: 'Changsha',
-      city_ascii: 'Changsha',
-      lat: 28.2,
-      lng: 112.97,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Hunan',
-      capital: 'admin',
-      population: 2604000,
-      id: 1156961497,
-    },
-    {
-      city: 'Dakar',
-      city_ascii: 'Dakar',
-      lat: 14.7158,
-      lng: -17.4731,
-      country: 'Senegal',
-      iso2: 'SN',
-      iso3: 'SEN',
-      admin_name: 'Dakar',
-      capital: 'primary',
-      population: 2604000,
-      id: 1686604760,
-    },
-    {
-      city: 'İzmir',
-      city_ascii: 'Izmir',
-      lat: 38.4361,
-      lng: 27.1518,
-      country: 'Turkey',
-      iso2: 'TR',
-      iso3: 'TUR',
-      admin_name: 'İzmir',
-      capital: 'admin',
-      population: 2587000,
-      id: 1792725579,
-    },
-    {
-      city: 'Xiangtan',
-      city_ascii: 'Xiangtan',
-      lat: 27.8504,
-      lng: 112.9,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Hunan',
-      capital: '',
-      population: 2586948,
-      id: 1156195684,
-    },
-    {
-      city: 'Lanzhou',
-      city_ascii: 'Lanzhou',
-      lat: 36.056,
-      lng: 103.792,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Gansu',
-      capital: 'admin',
-      population: 2561000,
-      id: 1156280566,
-    },
-    {
-      city: 'Incheon',
-      city_ascii: 'Incheon',
-      lat: 37.4761,
-      lng: 126.6422,
-      country: 'Korea, South',
-      iso2: 'KR',
-      iso3: 'KOR',
-      admin_name: 'Incheon',
-      capital: 'admin',
-      population: 2550000,
-      id: 1410160313,
-    },
-    {
-      city: 'Sapporo',
-      city_ascii: 'Sapporo',
-      lat: 43.075,
-      lng: 141.34,
-      country: 'Japan',
-      iso2: 'JP',
-      iso3: 'JPN',
-      admin_name: 'Hokkaidō',
-      capital: 'admin',
-      population: 2544000,
-      id: 1392000195,
-    },
-    {
-      city: 'Xiamen',
-      city_ascii: 'Xiamen',
-      lat: 24.45,
-      lng: 118.08,
-      country: 'China',
-      iso2: 'CN',
-      iso3: 'CHN',
-      admin_name: 'Fujian',
-      capital: 'minor',
-      population: 2519000,
-      id: 1156212809,
-    },
-    {
-      city: 'Guayaquil',
-      city_ascii: 'Guayaquil',
-      lat: -2.22,
-      lng: -79.92,
-      country: 'Ecuador',
-      iso2: 'EC',
-      iso3: 'ECU',
-      admin_name: 'Guayas',
-      capital: 'admin',
-      population: 2514000,
-      id: 1218802178,
-    },
-    {
-      city: 'George Town',
-      city_ascii: 'George Town',
-      lat: 5.4136,
-      lng: 100.3294,
-      country: 'Malaysia',
-      iso2: 'MY',
-      iso3: 'MYS',
-      admin_name: 'Pulau Pinang',
-      capital: 'admin',
-      population: 2500000,
-      id: 1458236750,
-    },
-    {
-      city: 'Mashhad',
-      city_ascii: 'Mashhad',
-      lat: 36.27,
-      lng: 59.57,
-      country: 'Iran',
-      iso2: 'IR',
-      iso3: 'IRN',
-      admin_name: 'Khorāsān-e Raẕavī',
-      capital: 'admin',
-      population: 2469000,
-      id: 1364123206,
-    },
-    {
-      city: 'Damascus',
-      city_ascii: 'Damascus',
-      lat: 33.5,
-      lng: 36.3,
-      country: 'Syria',
-      iso2: 'SY',
-      iso3: 'SYR',
-      admin_name: 'Dimashq',
-      capital: 'primary',
-      population: 2466000,
-      id: 1760685964,
-    },
-  ];
-  if (termSearch === '') {
-    return [];
-  }
-  return list;
-}
-
-export default CitiesData;
